+++ core/src/test/java/com/graphhopper/util/InstructionListTest.java
@@ -55,8 +55,8 @@ public class InstructionListTest {
     @BeforeEach
     public void setUp() {
         speedEnc = new DecimalEncodedValueImpl("speed", 5, 5, true);
-        carManager = EncodingManager.start().add(speedEnc).add(Roundabout.create())
-                .add(MaxSpeed.create()).add(RoadClass.create()).add(RoadClassLink.create()).build();
+        carManager = EncodingManager.start().add(speedEnc).add(Roundabout.create()).add(VehicleAccess.create("car")).
+                add(MaxSpeed.create()).add(RoadClass.create()).add(RoadClassLink.create()).build();
     }
 
     private static List<String> getTurnDescriptions(InstructionList instructionList) {
@@ -299,8 +299,9 @@ public class InstructionListTest {
     @Test
     public void testNoInstructionIfSlightTurnAndAlternativeIsSharp3() {
         DecimalEncodedValue speedEnc = new DecimalEncodedValueImpl("speed", 4, 2, true);
-        EncodingManager tmpEM = new EncodingManager.Builder().add(speedEnc).add(RoadClass.create())
-                .add(RoadClassLink.create()).add(Roundabout.create()).add(MaxSpeed.create()).build();
+        EncodingManager tmpEM = new EncodingManager.Builder().add(speedEnc).add(RoadClass.create()).
+                add(VehicleAccess.create("car")).add(RoadClassLink.create()).add(Roundabout.create()).
+                add(MaxSpeed.create()).build();
         EnumEncodedValue<RoadClass> rcEV = tmpEM.getEnumEncodedValue(RoadClass.KEY, RoadClass.class);
         BaseGraph g = new BaseGraph.Builder(tmpEM).create();
         // real world example: https://graphhopper.com/maps/?point=48.411549,15.599567&point=48.411663%2C15.600527&profile=bike
@@ -338,7 +339,9 @@ public class InstructionListTest {
     @Test
     public void testInstructionIfTurn() {
         DecimalEncodedValue speedEnc = new DecimalEncodedValueImpl("speed", 4, 2, true);
-        EncodingManager tmpEM = new EncodingManager.Builder().add(speedEnc).add(RoadClass.create()).add(RoadClassLink.create()).add(Roundabout.create()).add(MaxSpeed.create()).build();
+        EncodingManager tmpEM = new EncodingManager.Builder().add(speedEnc).add(RoadClass.create()).
+                add(VehicleAccess.create("car")).add(RoadClassLink.create()).add(Roundabout.create()).
+                add(MaxSpeed.create()).build();
         EnumEncodedValue<RoadClass> rcEV = tmpEM.getEnumEncodedValue(RoadClass.KEY, RoadClass.class);
         BaseGraph g = new BaseGraph.Builder(tmpEM).create();
         // real world example: https://graphhopper.com/maps/?point=48.412169%2C15.604888&point=48.412251%2C15.60543&profile=bike
@@ -375,8 +378,9 @@ public class InstructionListTest {
     @Test
     public void testInstructionIfSlightTurn() {
         DecimalEncodedValue speedEnc = new DecimalEncodedValueImpl("speed", 4, 1, false);
-        EncodingManager tmpEM = new EncodingManager.Builder().add(speedEnc)
-                .add(Roundabout.create()).add(RoadClass.create()).add(RoadClassLink.create()).add(MaxSpeed.create()).build();
+        EncodingManager tmpEM = new EncodingManager.Builder().add(speedEnc).add(Roundabout.create()).
+                add(VehicleAccess.create("car")).add(RoadClass.create()).add(RoadClassLink.create()).
+                add(MaxSpeed.create()).build();
         BaseGraph g = new BaseGraph.Builder(tmpEM).create();
         // real world example: https://graphhopper.com/maps/?point=43.729379,7.417697&point=43.729798,7.417263&profile=foot
         // From 4 to 3 and 4 to 1
@@ -428,8 +432,9 @@ public class InstructionListTest {
     public void testInstructionWithHighlyCustomProfileWithRoadsBase() {
         BooleanEncodedValue roadsAccessEnc = new SimpleBooleanEncodedValue("access", true);
         DecimalEncodedValue roadsSpeedEnc = new DecimalEncodedValueImpl("speed", 7, 2, true);
-        EncodingManager tmpEM = EncodingManager.start().add(roadsAccessEnc).add(roadsSpeedEnc)
-                .add(RoadClass.create()).add(Roundabout.create()).add(RoadClassLink.create()).add(MaxSpeed.create()).build();
+        EncodingManager tmpEM = EncodingManager.start().add(roadsAccessEnc).add(roadsSpeedEnc).
+                add(RoadClass.create()).add(Roundabout.create()).add(RoadClassLink.create()).
+                add(MaxSpeed.create()).add(VehicleAccess.create("car")).build();
         EnumEncodedValue<RoadClass> rcEV = tmpEM.getEnumEncodedValue(RoadClass.KEY, RoadClass.class);
         BaseGraph g = new BaseGraph.Builder(tmpEM).create();
         // real world example: https://graphhopper.com/maps/?point=55.691214%2C12.57065&point=55.689957%2C12.570387
@@ -523,7 +528,7 @@ public class InstructionListTest {
         DecimalEncodedValue roadsSpeedEnc = new DecimalEncodedValueImpl("speed", 7, 2, true);
         EncodingManager tmpEM = EncodingManager.start().add(roadsSpeedEnc).
                 add(RoadClass.create()).add(Roundabout.create()).add(RoadClassLink.create()).
-                add(MaxSpeed.create()).add(Lanes.create()).build();
+                add(MaxSpeed.create()).add(Lanes.create()).add(VehicleAccess.create("car")).build();
         IntEncodedValue lanesEnc = tmpEM.getIntEncodedValue(Lanes.KEY);
         BaseGraph g = new BaseGraph.Builder(tmpEM).create();
         // real world example: https://graphhopper.com/maps/?point=43.626238%2C-79.715268&point=43.624647%2C-79.713204&profile=car
@@ -568,7 +573,7 @@ public class InstructionListTest {
     @Test
     public void testNotSplitWays() {
         DecimalEncodedValue roadsSpeedEnc = new DecimalEncodedValueImpl("speed", 7, 2, true);
-        EncodingManager tmpEM = EncodingManager.start().add(roadsSpeedEnc).
+        EncodingManager tmpEM = EncodingManager.start().add(roadsSpeedEnc).add(VehicleAccess.create("car")).
                 add(RoadClass.create()).add(Roundabout.create()).add(RoadClassLink.create()).
                 add(MaxSpeed.create()).add(Lanes.create()).build();
         IntEncodedValue lanesEnc = tmpEM.getIntEncodedValue(Lanes.KEY);